GasPvtThermal: move implementation to impl file
